# onboarding-maxxidata-server/server/api/controllers/type-members/controller.js
import ResponseHandler from '../../../common/ResponseHandler';
import NotFoundError from '../../../common/exceptions/NotFoundError';
import Service from './TypeMemberService';
import l from '../../../common/logger';

export class Controller {
  async get(req, res) {
    try {
      l.debug(`Request -> ${JSON.stringify(req.ip)}`);
      const model = await Service.all(req.query);
      return res.json(model);
    } catch (e) {
      ResponseHandler.error(res, e);
    }
  }

  async getById(req, res) {
    try {
      const model = await Service.byId(req.params.id);

      if (!model) {
        throw new NotFoundError(req);
      }

      return res.json(model);
    } catch (e) {
      ResponseHandler.error(res, e);
    }
  }

  async post(req, res) {
    try {
      Service.create(req.body).then((r) =>
      res.status(201).location(`/api/v1/type_members/${r.id}`).json(r));
      /* return res.status(201).json(model); */
    } catch (e) {
      ResponseHandler.error(res, e);
    }
  }

  async put(req, res) {
    try {
      const { id } = req.params;
      const { fields } = req.body;

      const model = await Service.updateById(id, {
        fields
      });

      return res.status(201).json(model);
    } catch (e) {
      ResponseHandler.error(res, e);
    }
  }

  async delete(req, res) {
    try {
      const { id } = req.params;
      const model = await Service.deleteById(id);

      return res.status(201).json(model);
    } catch (e) {
      ResponseHandler.error(res, e);
    }
  }
}

export default new Controller();