# onboarding-maxxidata-server/server/common/api.yml
swagger: "2.0"
info:
  version: 1.0.0
  title: myapp
  description: Onboarding MaxxiData App
basePath: /api/v1
...
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /type_members/{id}:
    get:
      tags:
        - TypeMembers
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the example to retrieve
          type: integer
      responses:
        200:
          description: Return the members type with the specified id
        404:
          description: member type not found
...